+++ changelog.md
@@ -1,5 +1,10 @@
 All notable changes to this project will be documented in this file.
 
+### Added: Работа над ошибками
+- Удалёно `_` для переменной `productService` из `ProductListComponent`.
+- Добавлены строковые значения для enum файл `src/app/task1/models/product.ts`
+- Удалён `src/app/task1/services/converter.service.ts`, добавлена `Pipe` => `YesNoPipe` 
+
 ### Added: Complete bullet 6 from task list
 - Added `CardListComponent`
 - Added `CardComponent`